+++ tests/0200_views/0210_controller.py
@@ -4,7 +4,31 @@ from canopy.factories import ControllerFactory, SlotFactory
 from canopy.utils.tests import html_pyquery
 
 
-def test_view_detail_detail(db, client, django_assert_num_queries):
+def test_view_get_disabled(db, client, django_assert_num_queries, settings):
+    """
+    When disabled a GET request receives either a specific page for disabled form
+    or a Http404.
+    """
+    controller = ControllerFactory(enabled=False)
+    SlotFactory(controller=controller, name="basic-text")
+    url = reverse("canopy:controller-form", kwargs={"slug": controller.slug})
+
+    with django_assert_num_queries(1):
+        response = client.get(url, follow=True)
+
+    assert response.redirect_chain == []
+    assert response.status_code == 200
+
+    settings.CANOPY_CONTROLLER_EXCEPTION_WHEN_DISABLED = True
+
+    with django_assert_num_queries(1):
+        response = client.get(url, follow=True)
+
+    assert response.redirect_chain == []
+    assert response.status_code == 404
+
+
+def test_view_get_enabled(db, client, django_assert_num_queries):
     """
     Controller form view should contains all controller slot and a proper form
     structure.
@@ -30,7 +54,53 @@ def test_view_detail_detail(db, client, django_assert_num_queries):
     assert names == ["csrfmiddlewaretoken", "basic-text", "submit-request-form"]
 
 
-def test_view_detail_post(db, client, django_assert_num_queries):
+def test_view_success_disabled(db, client, django_assert_num_queries, settings):
+    """
+    When disabled the success view returns either a specific page for disabled form
+    or a Http404.
+    """
+    controller = ControllerFactory(enabled=False)
+    SlotFactory(controller=controller, name="basic-text", required=True)
+
+    url = reverse("canopy:controller-success", kwargs={"slug": controller.slug})
+
+    with django_assert_num_queries(1):
+        response = client.get(url, follow=True)
+    assert response.redirect_chain == []
+    assert response.status_code == 200
+
+    settings.CANOPY_CONTROLLER_EXCEPTION_WHEN_DISABLED = True
+
+    with django_assert_num_queries(1):
+        response = client.get(url, follow=True)
+    assert response.redirect_chain == []
+    assert response.status_code == 404
+
+
+def test_view_post_disabled(db, client, django_assert_num_queries, settings):
+    """
+    When disabled a POST request receives either a specific page for disabled form
+    or a Http404.
+    """
+    controller = ControllerFactory(enabled=False)
+    SlotFactory(controller=controller, name="basic-text", required=True)
+
+    form_url = reverse("canopy:controller-form", kwargs={"slug": controller.slug})
+
+    with django_assert_num_queries(1):
+        response = client.post(form_url, data={}, follow=True)
+    assert response.redirect_chain == []
+    assert response.status_code == 200
+
+    settings.CANOPY_CONTROLLER_EXCEPTION_WHEN_DISABLED = True
+
+    with django_assert_num_queries(1):
+        response = client.post(form_url, data={}, follow=True)
+    assert response.redirect_chain == []
+    assert response.status_code == 404
+
+
+def test_view_post_enabled(db, client, django_assert_num_queries):
     """
     Controller form should redirect to success view when submitted data have been
     validated. And the success view should retrieve the related controller and entry
@@ -49,21 +119,8 @@ def test_view_detail_post(db, client, django_assert_num_queries):
     assert response.status_code == 200
 
     # With required data, form is valid and goes to success
-    # NOTE: This involves two suite of querysets, the ones from form view then the ones
-    # from the success view
-    # with django_assert_num_queries(16) as captured_queries:
     response = client.post(form_url, data={"basic-text": "Foo"}, follow=True)
 
-    # import json
-    # print(json.dumps(list(captured_queries), indent=4))
-    # selection_queries = [
-    #     q["sql"].replace('\"', "")
-    #     for q in captured_queries
-    #     if q["sql"].startswith("SELECT")
-    # ]
-    # print()
-    # print(json.dumps(selection_queries, indent=4))
-    # assert 1 == 42
     assert response.redirect_chain == [(success_url, 302),]
     assert response.status_code == 200
     # Success view has retrieved related Controller object